render read-only Role section in TeamView and remove standalone substitutions sidebar
