--- src/lib/components/sidebar/SubstitutionsSidebar.svelte
@@ -1,597 +1,0 @@
-<script lang="ts">
-	import type {
-		GridCharacter,
-		GridWeapon,
-		GridSummon,
-		Role,
-		Substitution
-	} from '$lib/types/api/party'
-	import { createQuery } from '@tanstack/svelte-query'
-	import { roleQueries } from '$lib/api/queries/role.queries'
-	import { partyQueries } from '$lib/api/queries/party.queries'
-	import {
-		useCreateSubstitution,
-		useUpdateSubstitution,
-		useDeleteSubstitution
-	} from '$lib/api/mutations/substitution.mutations'
-	import { useUpdateGridWeapon } from '$lib/api/mutations/grid.mutations'
-	import { useUpdateGridCharacter } from '$lib/api/mutations/grid.mutations'
-	import { useUpdateGridSummon } from '$lib/api/mutations/grid.mutations'
-	import { sidebar } from '$lib/stores/sidebar.svelte'
-	import SearchContent from '$lib/components/sidebar/SearchContent.svelte'
-	import DetailsSection from '$lib/components/sidebar/details/DetailsSection.svelte'
-	import CharacterTags from '$lib/components/tags/CharacterTags.svelte'
-	import { localizedName } from '$lib/utils/locale'
-	import {
-		getCharacterImage,
-		getWeaponImage,
-		getSummonImage,
-		getPlaceholder
-	} from '$lib/features/database/detail/image'
-	import { getWeaponFallbackImage, handleImageFallback, STYLE_SWAP_POSE } from '$lib/utils/images'
-	import Select from '$lib/components/ui/Select.svelte'
-	import Button from '$lib/components/ui/Button.svelte'
-	import Icon from '$lib/components/Icon.svelte'
-	import * as m from '$lib/paraglide/messages'
-	import type { AddItemResult } from '$lib/types/api/search'
-
-	interface Props {
-		type: 'weapon' | 'character' | 'summon'
-		item: GridCharacter | GridWeapon | GridSummon
-		editable?: boolean
-		partyId?: string
-		partyShortcode?: string
-	}
-
-	let { type, item, editable = false, partyId, partyShortcode }: Props = $props()
-
-	// The pane stores `item` as a snapshot at open time, so prop reads alone go
-	// stale after a substitution mutation invalidates the party cache. Subscribe
-	// to the live party query and re-resolve the grid item by id on every read.
-	const partyQuery = createQuery(() => ({
-		...partyQueries.byShortcode(partyShortcode ?? ''),
-		enabled: !!partyShortcode
-	}))
-
-	const liveItem = $derived.by((): GridCharacter | GridWeapon | GridSummon | undefined => {
-		const party = partyQuery.data
-		if (!party || !item.id) return undefined
-		const list =
-			type === 'character' ? party.characters : type === 'summon' ? party.summons : party.weapons
-		return list?.find((g) => g.id === item.id) as
-			| GridCharacter
-			| GridWeapon
-			| GridSummon
-			| undefined
-	})
-
-	// Read from the live item when available, fall back to the prop snapshot.
-	const effectiveItem = $derived(liveItem ?? item)
-
-	// Derive data from the effective item
-	let role = $derived((effectiveItem as GridWeapon).role as Role | undefined)
-	let note = $derived((effectiveItem as GridWeapon).substitutionNote as string | undefined)
-	let substitutions = $derived(
-		// Spread before sort: Array.prototype.sort mutates in place, and a $derived
-		// expression must not mutate state — Svelte 5 throws state_unsafe_mutation
-		// and downstream reactivity breaks (added substitutes don't appear).
-		[...(((effectiveItem as GridWeapon).substitutions ?? []) as Substitution[])].sort(
-			(a, b) => a.position - b.position
-		)
-	)
-
-	// Underlying entity ids of items already added as substitutes for this slot,
-	// so the search modal can disable them and prevent the duplicate post.
-	const existingSubstituteItemIds = $derived(
-		substitutions
-			.map(
-				(s) =>
-					s.gridCharacter?.character?.id ??
-					s.gridWeapon?.weapon?.id ??
-					s.gridSummon?.summon?.id ??
-					null
-			)
-			.filter((id): id is string => id !== null)
-	)
-
-	// Slot type mapping for roles query
-	function getSlotType(t: string): string {
-		if (t === 'character') return 'Character'
-		if (t === 'summon') return 'Summon'
-		return 'Weapon'
-	}
-
-	// Grid type mapping for API
-	function getGridType(t: string): string {
-		if (t === 'character') return 'GridCharacter'
-		if (t === 'summon') return 'GridSummon'
-		return 'GridWeapon'
-	}
-
-	// Roles query
-	const rolesQuery = createQuery(() => ({
-		...roleQueries.bySlotType(getSlotType(type)),
-		enabled: editable
-	}))
-
-	const roleOptions = $derived.by(() => {
-		const roles = (rolesQuery.data ?? []) as Role[]
-		const options: { value: string; label: string }[] = [
-			{ value: '', label: m.substitution_role_none() }
-		]
-		for (const r of roles) {
-			options.push({
-				value: r.id,
-				label: localizedName({ en: r.nameEn, ja: r.nameJp }) ?? r.nameEn
-			})
-		}
-		return options
-	})
-
-	// Local state for editable fields (writable derived syncs from item)
-	let selectedRoleId = $derived(role?.id ?? '')
-	let noteText = $derived(note ?? '')
-
-	// Mutations
-	const createSubstitution = useCreateSubstitution()
-	const updateSubstitution = useUpdateSubstitution()
-	const deleteSubstitution = useDeleteSubstitution()
-	const updateWeapon = useUpdateGridWeapon()
-	const updateCharacter = useUpdateGridCharacter()
-	const updateSummon = useUpdateGridSummon()
-
-	// Save role change
-	function handleRoleChange(value: string | undefined) {
-		if (!item.id || !partyShortcode) return
-		const roleId = value || null
-		const updates = { roleId } as Record<string, unknown>
-
-		if (type === 'weapon') {
-			updateWeapon.mutate({
-				id: String(item.id),
-				partyShortcode,
-				updates: updates as Partial<GridWeapon>
-			})
-		} else if (type === 'character') {
-			updateCharacter.mutate({
-				id: String(item.id),
-				partyShortcode,
-				updates: updates as Partial<GridCharacter>
-			})
-		} else {
-			updateSummon.mutate({
-				id: String(item.id),
-				partyShortcode,
-				updates: updates as Partial<GridSummon>
-			})
-		}
-	}
-
-	// Save note on blur
-	function handleNoteBlur() {
-		if (!item.id || !partyShortcode) return
-		if (noteText === (note ?? '')) return
-		const updates = { substitutionNote: noteText || null } as Record<string, unknown>
-
-		if (type === 'weapon') {
-			updateWeapon.mutate({
-				id: String(item.id),
-				partyShortcode,
-				updates: updates as Partial<GridWeapon>
-			})
-		} else if (type === 'character') {
-			updateCharacter.mutate({
-				id: String(item.id),
-				partyShortcode,
-				updates: updates as Partial<GridCharacter>
-			})
-		} else {
-			updateSummon.mutate({
-				id: String(item.id),
-				partyShortcode,
-				updates: updates as Partial<GridSummon>
-			})
-		}
-	}
-
-	// Add substitute via search
-	function handleAddSubstitute() {
-		if (!partyId || !partyShortcode || !item.id) return
-
-		const searchPaneId = `search-substitute-${item.id}`
-		if (sidebar.paneStack.panes.some((p) => p.id === searchPaneId)) return
-
-		sidebar.push({
-			id: searchPaneId,
-			title: m.substitution_add_prompt(),
-			component: SearchContent,
-			props: {
-				type,
-				canAddMore: true,
-				excludedIds: existingSubstituteItemIds,
-				onAddItems: (items: AddItemResult[]) => {
-					const addItem = items[0]
-					if (!addItem) return
-					createSubstitution.mutate(
-						{
-							partyId: partyId!,
-							partyShortcode: partyShortcode!,
-							gridType: getGridType(type),
-							gridId: String(item.id),
-							itemId: addItem.id
-						},
-						{
-							onSuccess: () => sidebar.pop()
-						}
-					)
-				}
-			},
-			onback: () => sidebar.pop()
-		})
-	}
-
-	// Delete substitute
-	function handleDelete(sub: Substitution) {
-		if (!partyId || !partyShortcode) return
-		deleteSubstitution.mutate({
-			id: sub.id,
-			partyId: partyId!,
-			partyShortcode: partyShortcode!
-		})
-	}
-
-	// Drag-and-drop reorder. The substitutions table has no unique constraint
-	// on position, so we can sequentially write each item's new index without
-	// any temporary-position dance.
-	let dragIndex = $state<number | null>(null)
-	let hoverIndex = $state<number | null>(null)
-
-	function onDragStart(e: DragEvent, index: number) {
-		dragIndex = index
-		if (e.dataTransfer) {
-			e.dataTransfer.effectAllowed = 'move'
-			e.dataTransfer.setData('text/plain', String(index))
-		}
-	}
-
-	function onDragOver(e: DragEvent, index: number) {
-		e.preventDefault()
-		hoverIndex = index
-		if (e.dataTransfer) e.dataTransfer.dropEffect = 'move'
-	}
-
-	function onDragLeave() {
-		hoverIndex = null
-	}
-
-	async function onDrop(e: DragEvent, dropIndex: number) {
-		e.preventDefault()
-		const from = dragIndex
-		dragIndex = null
-		hoverIndex = null
-		if (from === null || from === dropIndex) return
-		if (!partyId || !partyShortcode) return
-
-		const next = [...substitutions]
-		const [moved] = next.splice(from, 1)
-		if (!moved) return
-		next.splice(dropIndex, 0, moved)
-
-		try {
-			for (let i = 0; i < next.length; i++) {
-				const target = next[i]
-				if (!target) continue
-				if (target.position === i) continue
-				await updateSubstitution.mutateAsync({
-					id: target.id,
-					partyId: partyId!,
-					partyShortcode: partyShortcode!,
-					position: i
-				})
-			}
-		} catch (err) {
-			console.error('Failed to reorder substitutions:', err)
-		}
-	}
-
-	function getSubstituteName(sub: Substitution): string {
-		if (sub.gridCharacter) return localizedName(sub.gridCharacter.character?.name) ?? '\u2014'
-		if (sub.gridWeapon) return localizedName(sub.gridWeapon.weapon?.name) ?? '\u2014'
-		if (sub.gridSummon) return localizedName(sub.gridSummon.summon?.name) ?? '\u2014'
-		return '\u2014'
-	}
-
-	function getSubstituteImage(sub: Substitution): string {
-		if (sub.gridCharacter?.character) {
-			const c = sub.gridCharacter.character
-			return getCharacterImage(c.granblueId, 'square', c.styleSwap ? STYLE_SWAP_POSE : '01')
-		}
-		if (sub.gridWeapon?.weapon) {
-			const w = sub.gridWeapon.weapon
-			return getWeaponImage(w.granblueId, 'square', w.element === 0 ? 0 : undefined)
-		}
-		if (sub.gridSummon?.summon) {
-			return getSummonImage(sub.gridSummon.summon.granblueId, 'square')
-		}
-		return getPlaceholder(type, 'square')
-	}
-
-	function getSubstituteFallbackImage(sub: Substitution): string | undefined {
-		if (sub.gridWeapon?.weapon && sub.gridWeapon.weapon.element === 0) {
-			return getWeaponFallbackImage(sub.gridWeapon.weapon.granblueId, 'square')
-		}
-		return undefined
-	}
-</script>
-
-<div class="substitutions-sidebar">
-	<DetailsSection title={m.substitution_role()}>
-		{#if editable}
-			<Select
-				options={roleOptions}
-				bind:value={selectedRoleId}
-				onValueChange={handleRoleChange}
-				placeholder={m.substitution_role_none()}
-				contained
-				fullWidth
-			/>
-		{:else if role}
-			<p class="role-name">
-				{localizedName({ en: role.nameEn, ja: role.nameJp })}
-			</p>
-		{:else}
-			<p class="empty">{m.substitution_role_none()}</p>
-		{/if}
-	</DetailsSection>
-
-	<DetailsSection title={m.substitution_note()}>
-		{#if editable}
-			<textarea
-				class="note-input contained"
-				bind:value={noteText}
-				onblur={handleNoteBlur}
-				placeholder={m.substitution_note_placeholder()}
-				rows="3"
-			></textarea>
-		{:else if note}
-			<p class="note">{note}</p>
-		{:else}
-			<p class="empty">{m.substitution_note_placeholder()}</p>
-		{/if}
-	</DetailsSection>
-
-	<DetailsSection title={m.substitution_substitutes()}>
-		{#if substitutions.length === 0}
-			<p class="empty">{m.substitution_empty()}</p>
-		{:else}
-			<ol class="substitution-list">
-				{#each substitutions as sub, index (sub.id)}
-					{@const character = sub.gridCharacter?.character}
-					<li
-						class="substitution-item"
-						class:drop-target={hoverIndex === index && dragIndex !== null && dragIndex !== index}
-						draggable={editable ? 'true' : 'false'}
-						ondragstart={(e) => onDragStart(e, index)}
-						ondragover={(e) => onDragOver(e, index)}
-						ondragleave={onDragLeave}
-						ondrop={(e) => onDrop(e, index)}
-					>
-						{#if editable}
-							<span class="handle" aria-hidden="true">
-								<Icon name="grip-vertical" width={4} height={18} />
-							</span>
-						{/if}
-						<img
-							src={getSubstituteImage(sub)}
-							alt=""
-							class="thumb"
-							loading="lazy"
-							onerror={(e) => handleImageFallback(e, getSubstituteFallbackImage(sub))}
-						/>
-						<div class="info">
-							<span class="name">{getSubstituteName(sub)}</span>
-							{#if character}
-								<div class="meta">
-									<CharacterTags {character} />
-								</div>
-							{/if}
-						</div>
-						{#if editable}
-							<div class="actions">
-								<button
-									class="action-btn delete"
-									onclick={() => handleDelete(sub)}
-									title={m.substitution_remove()}
-								>
-									<Icon name="close" size={14} />
-								</button>
-							</div>
-						{/if}
-					</li>
-				{/each}
-			</ol>
-		{/if}
-
-		{#if editable && substitutions.length < 10}
-			<Button variant="ghost" size="small" fullWidth leftIcon="plus" onclick={handleAddSubstitute}>
-				{m.substitution_add()}
-			</Button>
-		{/if}
-	</DetailsSection>
-</div>
-
-<style lang="scss">
-	@use '$src/themes/spacing' as spacing;
-	@use '$src/themes/typography' as typography;
-
-	.substitutions-sidebar {
-		padding: spacing.$unit-2x;
-		display: flex;
-		flex-direction: column;
-		gap: spacing.$unit-2x + spacing.$unit-half;
-	}
-
-	.role-name {
-		font-size: typography.$font-regular;
-		font-weight: typography.$medium;
-		margin: 0;
-	}
-
-	.note {
-		font-size: typography.$font-small;
-		line-height: 1.5;
-		margin: 0;
-		white-space: pre-wrap;
-	}
-
-	.note-input {
-		width: 100%;
-		font-family: inherit;
-		font-size: typography.$font-small;
-		line-height: 1.5;
-		padding: spacing.$unit;
-		border: 2px solid transparent;
-		border-radius: spacing.$unit;
-		background: transparent;
-		color: var(--text-primary);
-		resize: vertical;
-		min-height: 60px;
-		box-sizing: border-box;
-
-		&::placeholder {
-			color: var(--text-tertiary);
-		}
-
-		&:focus {
-			outline: none;
-			border-color: var(--accent-blue);
-		}
-
-		&.contained {
-			background-color: var(--input-bound-bg);
-
-			&:hover:not(:disabled) {
-				background-color: var(--input-bound-bg-hover);
-			}
-		}
-	}
-
-	.empty {
-		font-size: typography.$font-small;
-		color: var(--text-tertiary);
-		margin: 0;
-	}
-
-	.substitution-list {
-		list-style: none;
-		padding: 0;
-		margin: 0;
-		display: flex;
-		flex-direction: column;
-		gap: spacing.$unit-half;
-	}
-
-	.substitution-item {
-		display: flex;
-		align-items: center;
-		gap: spacing.$unit;
-		font-size: typography.$font-regular;
-		padding: spacing.$unit;
-		border-radius: spacing.$unit;
-		transition:
-			background 0.15s ease,
-			outline-color 0.15s ease;
-		outline: 2px solid transparent;
-
-		&:hover {
-			background: var(--list-cell-bg-hover);
-		}
-
-		&.drop-target {
-			outline-color: var(--accent-blue);
-		}
-	}
-
-	.handle {
-		display: inline-flex;
-		align-items: center;
-		justify-content: center;
-		color: var(--text-tertiary);
-		cursor: grab;
-		flex-shrink: 0;
-
-		&:active {
-			cursor: grabbing;
-		}
-	}
-
-	.thumb {
-		width: 64px;
-		height: 64px;
-		object-fit: cover;
-		border-radius: spacing.$unit-half;
-		background: var(--placeholder-bg);
-		flex-shrink: 0;
-	}
-
-	.info {
-		flex: 1;
-		min-width: 0;
-		display: flex;
-		flex-direction: column;
-		gap: spacing.$unit-half;
-	}
-
-	.name {
-		min-width: 0;
-		overflow: hidden;
-		text-overflow: ellipsis;
-		white-space: nowrap;
-		color: var(--text-primary);
-	}
-
-	.meta {
-		display: flex;
-		flex-wrap: wrap;
-		align-items: center;
-		gap: spacing.$unit-half;
-	}
-
-	.actions {
-		display: flex;
-		align-items: center;
-		gap: 2px;
-		flex-shrink: 0;
-	}
-
-	.action-btn {
-		display: flex;
-		align-items: center;
-		justify-content: center;
-		width: 24px;
-		height: 24px;
-		padding: 0;
-		border: none;
-		border-radius: spacing.$unit-half;
-		background: transparent;
-		color: var(--text-tertiary);
-		cursor: pointer;
-		transition:
-			color 0.15s ease,
-			background 0.15s ease;
-
-		&:hover:not(:disabled) {
-			color: var(--text-primary);
-			background: var(--input-bound-bg-hover);
-		}
-
-		&:disabled {
-			opacity: 0.3;
-			cursor: not-allowed;
-		}
-
-		&.delete:hover:not(:disabled) {
-			color: var(--red);
-		}
-	}
-</style>
